Split worker processing logic.
